Moves Classes and Media INTO Subdirectories
