+++ lib/colors.py
